% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{add_comment}
\alias{add_comment}
\title{Add a comment to a task or project}
\usage{
add_comment(
  content,
  task_id = NULL,
  project_id = NULL,
  verbose = TRUE,
  token = get_todoist_api_token()
)
}
\arguments{
\item{content}{content of the comment}

\item{task_id}{id of the task (either task_id or project_id required)}

\item{project_id}{id of the project (either task_id or project_id required)}

\item{verbose}{boolean that make the function verbose}

\item{token}{todoist API token}
}
\value{
id of the new comment
}
\description{
Add a comment to a task or project
}
\examples{
\dontrun{
add_comment(content = "This is a comment", task_id = "12345")
add_comment(content = "Project comment", project_id = "67890")
}
}
