% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{add_responsible_to_task}
\alias{add_responsible_to_task}
\title{Add responsible to a task}
\usage{
add_responsible_to_task(
  project_id = get_project_id(project_name = project_name, token = token),
  project_name,
  responsible,
  task,
  verbose = FALSE,
  all_users = get_all_users(token = token),
  token = get_todoist_api_token()
)
}
\arguments{
\item{project_id}{id of the project}

\item{project_name}{name of the project}

\item{responsible}{add someone to this task with mail}

\item{task}{the full name of the task}

\item{verbose}{boolean that make the function verbose}

\item{all_users}{all_users}

\item{token}{todoist API token}
}
\value{
http request
}
\description{
Add responsible to a task
}
