% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{add_section}
\alias{add_section}
\title{Title}
\usage{
add_section(project_id, section_name, token = get_todoist_api_token())
}
\arguments{
\item{project_id}{}

\item{section_name}{}
}
\value{

}
\description{
Title
}
