% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{add_task_in_project}
\alias{add_task_in_project}
\title{Add task in project}
\usage{
add_task_in_project(
  project_id,
  task,
  try_again = 3,
  time_try_again = 3,
  verbose = TRUE,
  responsible = NULL,
  token = get_todoist_api_token(),
  les_taches = get_tasks(token = token)
)
}
\arguments{
\item{project_id}{id of a project}

\item{task}{a task}

\item{try_again}{start again the request}

\item{verbose}{make it talk}

\item{token}{token}

\item{add_responsible}{add responsible}
}
\value{
message
}
\description{
Add task in project
}
