% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{add_tasks_in_project_from_df}
\alias{add_tasks_in_project_from_df}
\title{Add tasks in project}
\usage{
add_tasks_in_project_from_df(
  project_id,
  tasks_list,
  verbose = FALSE,
  existing_tasks = get_tasks(token = token),
  token = get_todoist_api_token()
)
}
\arguments{
\item{project_id}{id of project}

\item{tasks_list}{list of tasks}

\item{verbose}{make it talk}

\item{existing_tasks}{existing tasks}

\item{token}{todoist API token}

\item{responsible}{add people in project}

\item{due}{due date}

\item{section_id}{section id}
}
\value{
id of project (character vector)
}
\description{
Add tasks in project
}
\seealso{
[add_task_in_project()]
}
