% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{add_user_in_project}
\alias{add_user_in_project}
\title{Add one user}
\usage{
add_user_in_project(
  project_id,
  mail,
  verbose = TRUE,
  token = get_todoist_api_token()
)
}
\arguments{
\item{project_id}{id of project}

\item{mail}{mail of the user}

\item{verbose}{make it talk}

\item{token}{token}
}
\description{
Add one user
}
