% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{archive_section}
\alias{archive_section}
\title{Archive a section}
\usage{
archive_section(section_id, verbose = TRUE, token = get_todoist_api_token())
}
\arguments{
\item{section_id}{id of the section}

\item{verbose}{boolean that make the function verbose}

\item{token}{todoist API token}
}
\value{
id of the archived section (invisible)
}
\description{
Archive a section
}
\examples{
\dontrun{
archive_section("12345")
}
}
