% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{ask_todoist_api_token}
\alias{ask_todoist_api_token}
\title{ask_todoist_api_token}
\usage{
ask_todoist_api_token(msg = "todoist api token")
}
\arguments{
\item{msg}{the message}
}
\description{
ask for the todoiste api token
}
