% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_token.R
\name{ask_todoist_app_api_token}
\alias{ask_todoist_app_api_token}
\title{Get the token}
\usage{
ask_todoist_app_api_token(
  app_id = rstudioapi::showPrompt("Name of app", ""),
  client_id = rstudioapi::showPrompt("Client ID of the app", message = "Client ID"),
  client_secret = rstudioapi::askForPassword(prompt = "Client Secret of app")
)
}
\arguments{
\item{app_id}{name of the app}

\item{client_id}{client id}

\item{client_secret}{client secret}
}
\value{
object of connection
}
\description{
Get the token
}
