% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{call_api}
\alias{call_api}
\title{Call the good version of API}
\usage{
call_api(..., url = "https://todoist.com/api/v8/sync")
}
\arguments{
\item{...}{any params of POST request}

\item{url}{url to call}
}
\value{
list
}
\description{
Call the good version of API
}
