% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{call_api_project_data}
\alias{call_api_project_data}
\title{Call project data}
\usage{
call_api_project_data(...)
}
\arguments{
\item{...}{any params of POST request}
}
\value{
list
}
\description{
Call project data
}
