% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{call_api_rest}
\alias{call_api_rest}
\title{Call the REST API}
\usage{
call_api_rest(endpoint, token = get_todoist_api_token(), ...)
}
\arguments{
\item{endpoint}{API endpoint (e.g., "projects", "tasks")}

\item{token}{todoist API token}

\item{...}{query parameters}
}
\value{
list
}
\description{
Call the REST API
}
