% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{delete_filter}
\alias{delete_filter}
\title{Delete a filter}
\usage{
delete_filter(
  filter_id = get_filter_id(filter_name = filter_name, token = token),
  filter_name,
  verbose = TRUE,
  token = get_todoist_api_token()
)
}
\arguments{
\item{filter_id}{id of the filter}

\item{filter_name}{name of the filter (for lookup if filter_id not provided)}

\item{verbose}{boolean that make the function verbose}

\item{token}{todoist API token}
}
\value{
NULL (invisible)
}
\description{
Delete a filter
}
\examples{
\dontrun{
delete_filter(filter_name = "Urgent Today")
}
}
