% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{delete_label}
\alias{delete_label}
\title{Delete a label}
\usage{
delete_label(
  label_id = get_label_id(label_name = label_name, token = token, create = FALSE),
  label_name,
  verbose = TRUE,
  token = get_todoist_api_token()
)
}
\arguments{
\item{label_id}{id of the label}

\item{label_name}{name of the label (for lookup if label_id not provided)}

\item{verbose}{boolean that make the function verbose}

\item{token}{todoist API token}
}
\value{
NULL (invisible)
}
\description{
Delete a label
}
\examples{
\dontrun{
delete_label(label_name = "urgent")
}
}
