\name{delete_project}
\alias{delete_project}
\title{delete_project}
\usage{
  delete_project(project_name,
    token = getOption("TodoistToken"))
}
\arguments{
  \item{project_name}{<what param does>}

  \item{token}{= getOption('TodoistToken') <what param
  does>}
}
\description{
  Deletes an existing Todoist project
}
\examples{
\dontrun{
delete_project('unwanted_project')
}
}

