% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reminders.R
\name{delete_reminder}
\alias{delete_reminder}
\title{Delete a reminder}
\usage{
delete_reminder(reminder_id, verbose = TRUE, token = get_todoist_api_token())
}
\arguments{
\item{reminder_id}{id of the reminder}

\item{verbose}{boolean that make the function verbose}

\item{token}{todoist API token}
}
\value{
NULL (invisible)
}
\description{
Delete a reminder
}
\examples{
\dontrun{
delete_reminder("12345")
}
}
