% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{delete_task}
\alias{delete_task}
\title{Delete a task}
\usage{
delete_task(task_id, verbose = TRUE, token = get_todoist_api_token())
}
\arguments{
\item{task_id}{id of the task to delete}

\item{verbose}{boolean that make the function verbose}

\item{token}{todoist API token}
}
\value{
NULL (invisible)
}
\description{
Delete a task
}
\examples{
\dontrun{
delete_task("12345")
}
}
