% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploads.R
\name{delete_upload}
\alias{delete_upload}
\title{Delete an uploaded file}
\usage{
delete_upload(file_url, verbose = TRUE, token = get_todoist_api_token())
}
\arguments{
\item{file_url}{URL of the file to delete}

\item{verbose}{boolean that make the function verbose}

\item{token}{todoist API token}
}
\value{
NULL (invisible)
}
\description{
Delete an uploaded file
}
\examples{
\dontrun{
delete_upload("https://...")
}
}
