% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backups.R
\name{download_backup}
\alias{download_backup}
\title{Download a backup}
\usage{
download_backup(
  version,
  output_file,
  verbose = TRUE,
  token = get_todoist_api_token()
)
}
\arguments{
\item{version}{version of the backup to download (from get_backups())}

\item{output_file}{path where to save the backup file}

\item{verbose}{boolean that make the function verbose}

\item{token}{todoist API token}
}
\value{
path to the saved file (invisible)
}
\description{
Download a backup
}
\examples{
\dontrun{
backups <- get_backups()
download_backup(backups$version[1], "backup.zip")
}
}
