% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{export_template}
\alias{export_template}
\title{Export a project as a template}
\usage{
export_template(
  project_id = get_project_id(project_name = project_name, token = token, create = FALSE),
  project_name,
  output_file,
  verbose = TRUE,
  token = get_todoist_api_token()
)
}
\arguments{
\item{project_id}{id of the project to export}

\item{project_name}{name of the project (for lookup if project_id not provided)}

\item{output_file}{path where to save the template file}

\item{verbose}{boolean that make the function verbose}

\item{token}{todoist API token}
}
\value{
path to the saved file (invisible)
}
\description{
Export a project as a template
}
\examples{
\dontrun{
export_template(project_name = "my_proj", output_file = "template.csv")
}
}
