% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{get_all_filters}
\alias{get_all_filters}
\title{Get all filters}
\usage{
get_all_filters(token = get_todoist_api_token())
}
\arguments{
\item{token}{todoist API token}
}
\value{
tibble of all filters
}
\description{
Get all filters
}
\examples{
\dontrun{
get_all_filters()
}
}
