% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{get_all_labels}
\alias{get_all_labels}
\title{Get all labels}
\usage{
get_all_labels(token = get_todoist_api_token())
}
\arguments{
\item{token}{todoist API token}
}
\value{
tibble of all labels
}
\description{
Get all labels
}
\examples{
\dontrun{
get_all_labels()
}
}
