% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{get_all_sections}
\alias{get_all_sections}
\title{Get all sections}
\usage{
get_all_sections(token = get_todoist_api_token())
}
\arguments{
\item{token}{todoist API token}
}
\value{
tibble of all sections
}
\description{
Get all sections
}
\examples{
\dontrun{
get_all_sections()
}
}
