% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backups.R
\name{get_backups}
\alias{get_backups}
\title{Get list of backups}
\usage{
get_backups(token = get_todoist_api_token())
}
\arguments{
\item{token}{todoist API token}
}
\value{
tibble of available backups
}
\description{
Get list of backups
}
\examples{
\dontrun{
get_backups()
}
}
