% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{get_comment}
\alias{get_comment}
\title{Get a single comment by ID}
\usage{
get_comment(comment_id, token = get_todoist_api_token())
}
\arguments{
\item{comment_id}{id of the comment}

\item{token}{todoist API token}
}
\value{
list with comment details
}
\description{
Get a single comment by ID
}
\examples{
\dontrun{
get_comment("12345")
}
}
