% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{get_comments}
\alias{get_comments}
\title{Get comments}
\usage{
get_comments(
  task_id = NULL,
  project_id = NULL,
  token = get_todoist_api_token()
)
}
\arguments{
\item{task_id}{id of the task (either task_id or project_id required)}

\item{project_id}{id of the project (either task_id or project_id required)}

\item{token}{todoist API token}
}
\value{
tibble of comments
}
\description{
Get comments
}
\examples{
\dontrun{
get_comments(task_id = "12345")
get_comments(project_id = "67890")
}
}
