% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{get_filter}
\alias{get_filter}
\title{Get a single filter by ID}
\usage{
get_filter(filter_id, token = get_todoist_api_token())
}
\arguments{
\item{filter_id}{id of the filter}

\item{token}{todoist API token}
}
\value{
list with filter details
}
\description{
Get a single filter by ID
}
\examples{
\dontrun{
get_filter("12345")
}
}
