% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{get_id_project}
\alias{get_id_project}
\title{Get id of project}
\usage{
get_id_project(object, project_name)
}
\arguments{
\item{object}{result of get_projects}

\item{project_name}{name of the project}
}
\value{
id of project (character vector)
}
\description{
This function gives you the id of a project by name, which is useful for adding tasks or people to the project.
}
\examples{
\dontrun{
get_projects() \%>\%
    get_id_project("test")
}
}
