% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{get_id_section}
\alias{get_id_section}
\title{id section}
\usage{
get_id_section(project_id, section_name, token = get_todoist_api_token())
}
\arguments{
\item{project_id}{num of the project id}

\item{section_name}{name of the section}
}
\description{
id section
}
