% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{get_label_id}
\alias{get_label_id}
\title{Get label id by name}
\usage{
get_label_id(
  label_name,
  all_labels = get_all_labels(token = token),
  token = get_todoist_api_token(),
  create = TRUE
)
}
\arguments{
\item{label_name}{name of the label}

\item{all_labels}{result of get_all_labels (optional)}

\item{token}{todoist API token}

\item{create}{boolean create label if needed}
}
\value{
id of the label
}
\description{
Get label id by name
}
\examples{
\dontrun{
get_label_id("urgent")
}
}
