% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_productivity_stats}
\alias{get_productivity_stats}
\title{Get productivity stats}
\usage{
get_productivity_stats(token = get_todoist_api_token())
}
\arguments{
\item{token}{todoist API token}
}
\value{
list with productivity statistics
}
\description{
Get productivity stats
}
\examples{
\dontrun{
get_productivity_stats()
}
}
