% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{get_project}
\alias{get_project}
\title{Get a single project by ID}
\usage{
get_project(project_id, token = get_todoist_api_token())
}
\arguments{
\item{project_id}{id of the project}

\item{token}{todoist API token}
}
\value{
list with project details
}
\description{
Get a single project by ID
}
\examples{
\dontrun{
get_project("12345")
}
}
