% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{get_project_id}
\alias{get_project_id}
\title{Get id of project}
\usage{
get_project_id(
  project_name,
  all_projects = get_all_projects(token = token),
  token = get_todoist_api_token(),
  create = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{project_name}{name of the project}

\item{all_projects}{result of \link[rtodoist]{get_all_projects}}

\item{token}{todoist API token}

\item{create}{boolean create project if needed}

\item{verbose}{boolean that make the function verbose}
}
\value{
id of project (character vector)
}
\description{
This function gives you the id of a project by name, which is useful for adding tasks or people to the project.
}
\examples{
\dontrun{
get_all_projects() \%>\%
    get_project_id("test")
}
}
