% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{get_shared_labels}
\alias{get_shared_labels}
\title{Get shared labels}
\usage{
get_shared_labels(token = get_todoist_api_token())
}
\arguments{
\item{token}{todoist API token}
}
\value{
tibble of shared labels
}
\description{
Get shared labels
}
\examples{
\dontrun{
get_shared_labels()
}
}
