% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{get_task}
\alias{get_task}
\title{Get a single task by ID}
\usage{
get_task(task_id, token = get_todoist_api_token())
}
\arguments{
\item{task_id}{id of the task}

\item{token}{todoist API token}
}
\value{
list with task details
}
\description{
Get a single task by ID
}
\examples{
\dontrun{
get_task("12345")
}
}
