% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_user_id}
\alias{get_user_id}
\title{Get user id}
\usage{
get_user_id(mail, token = get_todoist_api_token())
}
\arguments{
\item{mail}{mail of the person}

\item{token}{token}
}
\value{
id of user
}
\description{
Get user id
}
