% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_user_info}
\alias{get_user_info}
\title{Get current user info}
\usage{
get_user_info(token = get_todoist_api_token())
}
\arguments{
\item{token}{todoist API token}
}
\value{
list with user information
}
\description{
Get current user info
}
\examples{
\dontrun{
get_user_info()
}
}
