% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_users}
\alias{get_users}
\title{Get users}
\usage{
get_users(token = get_todoist_api_token())
}
\arguments{
\item{token}{token}
}
\value{
tibble of users
}
\description{
Get a tibble with emails and ids of users
}
\examples{
\dontrun{
get_users()
}
}
