% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_users_id}
\alias{get_users_id}
\title{Get users id}
\usage{
get_users_id(
  mails,
  all_users = get_all_users(token = token),
  token = get_todoist_api_token()
)
}
\arguments{
\item{mails}{mails of the person}

\item{all_users}{all_users}

\item{token}{token}
}
\value{
id of users
}
\description{
Get users id
}
