% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_users_in_project}
\alias{get_users_in_project}
\title{Get users in projects}
\usage{
get_users_in_project(project_id, token = get_todoist_api_token())
}
\arguments{
\item{project_id}{id of project}

\item{token}{token}
}
\value{
dataframe of users in projects
}
\description{
Get users in projects
}
