% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_presta.R
\name{init_forma_admin}
\alias{init_forma_admin}
\title{Title}
\usage{
init_forma_admin(
  project_id,
  tasks_list = list("recevoir nom/prenom et mail des stagiaires",
    "envoyer test de préévaluation", "mettre les echeances aux taches suivantes",
    "envoyer convention", "vérifier retour convention",
    "s'assurer que la logistique du formateur est OK",
    "envoyer convocation aux stagiaires", "upload feuille emmargement",
    "envoyer facture", "verifier que la facture est bien dans le bon projet axonaut",
    "envoyer attestation", "envoyer evaluation a chaud",
    "plannifier l'evaluation a froid", "verifie retour eval a cahud",     
    "verifie retour eval à froid", "mail retour de qualité + audt besoin"),
  try_again = 3,
  time_try_again = 3,
  verbose = TRUE,
  responsible = NULL,
  token = get_todoist_api_token()
)
}
\arguments{
\item{token}{}
}
\value{

}
\description{
Title
}
