% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_presta.R
\name{init_forma_manager}
\alias{init_forma_manager}
\title{Init formation manager}
\usage{
init_forma_manager(
  project_id,
  tasks_list = list("Envoi du devis + contenu",
    "Assigner au formateur les taches du formateur", "Etre Pay\\\\u00E9"),
  try_again = 3,
  time_try_again = 3,
  verbose = TRUE,
  responsible = NULL,
  token = get_todoist_api_token()
)
}
\arguments{
\item{project_id}{id of project}

\item{tasks_list}{lists of tasks}

\item{try_again}{try again the http request}

\item{time_try_again}{number of tries}

\item{verbose}{make the function verbose}

\item{responsible}{add people in project}

\item{token}{token}
}
\description{
Init formation manager
}
