% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_presta.R
\name{init_inter}
\alias{init_inter}
\title{Init project with inter template}
\usage{
init_inter(
  project_id,
  tasks_list = list("S'assurer d'avoir un nom de projet coherent avec Slack",
    "S'assigner et mettre des dates sur certaines taches pour pas les oublier",
    "Envoyer les questionnaires", "Envoi du devis + contenu",
    "Envoi des conventions de formation", "Envoi des convocations",
    "Creer le contenu de la formation", "Reserver salle",
    "Envoi des liens d'installations", "Imprimer des feuilles d'emargement",
    "S'assurer d'avoir son materiel complet", "Reserver Resto",
    "Remplir ses notes de frais", "Scanner la feuille d'emargement",     
    "Enregistrer son temps de formation dans Toggl",
    "Envoyer le questionnaire de satisfaction",
    "Envoyer questionnaire de satisfaction a froid", "Facturer", "Paye"),
  try_again = 3,
  time_try_again = 3,
  verbose = TRUE,
  token = get_todoist_api_token()
)
}
\arguments{
\item{project_id}{project id}

\item{tasks_list}{list pf tasks}

\item{try_again}{start again the request}

\item{verbose}{make it talk}

\item{token}{token}
}
\description{
Init project with inter template
}
