% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_presta.R
\name{init_presta_admin}
\alias{init_presta_admin}
\title{Title}
\usage{
init_presta_admin(
  project_id,
  tasks_list = list("Facturer", "Etre Payé"),
  try_again = 3,
  time_try_again = 3,
  verbose = TRUE,
  responsible = NULL,
  token = get_todoist_api_token()
)
}
\arguments{
\item{token}{}
}
\value{

}
\description{
Title
}
