% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_presta.R
\name{init_presta_manager}
\alias{init_presta_manager}
\title{Title}
\usage{
init_presta_manager(
  project_id,
  tasks_list = list("Proposition - Devis", "Gestion projet",
    "S'assurer d'avoir un nom de projet coherent avec Slack",
    "S'assigner et mettre des dates sur certaines taches pour pas les oublier",
    "Rediger la reference de la mission dans {reference}"),
  try_again = 3,
  time_try_again = 3,
  verbose = TRUE,
  responsible = NULL,
  token = get_todoist_api_token()
)
}
\arguments{
\item{token}{}
}
\value{

}
\description{
Title
}
