% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{invite_to_workspace}
\alias{invite_to_workspace}
\title{Invite user to workspace}
\usage{
invite_to_workspace(
  workspace_id,
  email,
  role = "member",
  verbose = TRUE,
  token = get_todoist_api_token()
)
}
\arguments{
\item{workspace_id}{id of the workspace}

\item{email}{email of the user to invite}

\item{role}{role for the user (e.g., "member", "admin")}

\item{verbose}{boolean that make the function verbose}

\item{token}{todoist API token}
}
\value{
NULL (invisible)
}
\description{
Invite user to workspace
}
\examples{
\dontrun{
invite_to_workspace("12345", "user@example.com")
}
}
