% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{leave_workspace}
\alias{leave_workspace}
\title{Leave a workspace}
\usage{
leave_workspace(workspace_id, verbose = TRUE, token = get_todoist_api_token())
}
\arguments{
\item{workspace_id}{id of the workspace}

\item{verbose}{boolean that make the function verbose}

\item{token}{todoist API token}
}
\value{
NULL (invisible)
}
\description{
Leave a workspace
}
\examples{
\dontrun{
leave_workspace("12345")
}
}
