% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{move_section}
\alias{move_section}
\title{Move a section to another project}
\usage{
move_section(
  section_id,
  project_id,
  verbose = TRUE,
  token = get_todoist_api_token()
)
}
\arguments{
\item{section_id}{id of the section}

\item{project_id}{target project id}

\item{verbose}{boolean that make the function verbose}

\item{token}{todoist API token}
}
\value{
id of the moved section (invisible)
}
\description{
Move a section to another project
}
\examples{
\dontrun{
move_section("12345", project_id = "67890")
}
}
