% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_todoist_website.R
\name{open_todoist_website_profile}
\alias{open_todoist_website_profile}
\title{open todoist website}
\usage{
open_todoist_website_profile(verbose = TRUE)
}
\arguments{
\item{verbose}{make it talk}
}
\value{
open a webpage
}
\description{
open todoist website
}
