% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{reject_invitation}
\alias{reject_invitation}
\title{Reject a project invitation}
\usage{
reject_invitation(
  invitation_id,
  invitation_secret,
  verbose = TRUE,
  token = get_todoist_api_token()
)
}
\arguments{
\item{invitation_id}{id of the invitation}

\item{invitation_secret}{secret of the invitation}

\item{verbose}{boolean that make the function verbose}

\item{token}{todoist API token}
}
\value{
NULL (invisible)
}
\description{
Reject a project invitation
}
\examples{
\dontrun{
reject_invitation("12345", "secret123")
}
}
