% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{remove_shared_label}
\alias{remove_shared_label}
\title{Remove a shared label}
\usage{
remove_shared_label(name, verbose = TRUE, token = get_todoist_api_token())
}
\arguments{
\item{name}{name of the shared label to remove}

\item{verbose}{boolean that make the function verbose}

\item{token}{todoist API token}
}
\value{
NULL (invisible)
}
\description{
Remove a shared label
}
\examples{
\dontrun{
remove_shared_label("label_name")
}
}
