% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{rename_shared_label}
\alias{rename_shared_label}
\title{Rename a shared label}
\usage{
rename_shared_label(
  old_name,
  new_name,
  verbose = TRUE,
  token = get_todoist_api_token()
)
}
\arguments{
\item{old_name}{current name of the shared label}

\item{new_name}{new name for the shared label}

\item{verbose}{boolean that make the function verbose}

\item{token}{todoist API token}
}
\value{
NULL (invisible)
}
\description{
Rename a shared label
}
\examples{
\dontrun{
rename_shared_label("old_name", "new_name")
}
}
