% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtodoist-package.R
\docType{package}
\name{rtodoist-package}
\alias{rtodoist}
\alias{rtodoist-package}
\title{rtodoist: R interface to todoist.com}
\description{
Manage projects and tasks on Todoist via R.
}
\keyword{internal}
