% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{set_todoist_api_token}
\alias{set_todoist_api_token}
\title{set_todoist_api_token}
\usage{
set_todoist_api_token(token)
}
\arguments{
\item{token}{todoist api token}
}
\description{
set the todoiste api token
}
