% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{tasks}
\alias{tasks}
\alias{tsk_get_all}
\alias{tsk_get_by_id}
\alias{tsk_add}
\alias{tsk_update_by_id}
\alias{tsk_close}
\alias{tsk_reopen}
\alias{tsk_delete}
\title{Modify tasks on the todoist list}
\usage{
tsk_get_all(project_id = NULL, label_id = NULL, filter = NULL,
  lang = NULL, token = use_token())

tsk_get_by_id(id, token = use_token())

tsk_add(content = NULL, project_id = NULL, parent = NULL,
  order = NULL, label_ids = NULL, priority = NULL,
  due_string = NULL, due_date = NULL, due_datetime = NULL,
  due_lang = NULL, token = use_token())

tsk_update_by_id(id, content = NULL, label_ids = NULL,
  priority = NULL, due_string = NULL, due_date = NULL,
  due_datetime = NULL, due_lang = NULL, token = use_token())

tsk_close(id, token = use_token())

tsk_reopen(id, token = use_token())

tsk_delete(id, token = use_token())
}
\arguments{
\item{project_id}{Task project id. If not set, task is put to user's inbox.}

\item{content}{Task content.}

\item{parent}{Parent task id.}

\item{order}{Non-zero integer value to sort tasks under the same parent.}

\item{label_ids}{A vector of strings. Ids of labels associated with the task.}

\item{priority}{An integer ranging from priority from 1 (normal) to 4 (urgent)}

\item{due_string}{Human defined task due date (e.g. "next Monday", "Tomorrow"). Value is set using local (not UTC) tie.}

\item{due_date}{Specific date in YYYY-MM-DD format relative to user's timezone}

\item{due_datetime}{Specific data and time in RFC3339 format in UTC}

\item{due_lang}{2-letter code specifying language in case `due_string` is not written in English.}
}
\description{
Only one of the `due_*` can be used at the same time (except `due_lang`).
}
