% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todoist-package.R
\docType{package}
\name{todoist-package}
\alias{todoist}
\alias{todoist-package}
\title{todoist: Create and Manage Todolist using 'Todoist.com' API}
\description{
'Todoist' <https://todoist.com/> provides an online task manager service for teams.
    Functions are provided that make it possible to interact with the 'Todoist' platform API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ThinkR-open/todoist}
  \item Report bugs at \url{https://github.com/ThinkR-open/todoist/issues}
}

}
\author{
\strong{Maintainer}: Cervan Girard \email{cervan@thinkr.fr} (\href{https://orcid.org/0000-0002-4816-4624}{ORCID})

Authors:
\itemize{
  \item Vincent Guyader \email{vincent@thinkr.fr} (\href{https://orcid.org/0000-0003-0671-9270}{ORCID})
}

Other contributors:
\itemize{
  \item ThinkR [copyright holder, funder]
}

}
\keyword{internal}
