% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{unarchive_section}
\alias{unarchive_section}
\title{Unarchive a section}
\usage{
unarchive_section(section_id, verbose = TRUE, token = get_todoist_api_token())
}
\arguments{
\item{section_id}{id of the section}

\item{verbose}{boolean that make the function verbose}

\item{token}{todoist API token}
}
\value{
id of the unarchived section (invisible)
}
\description{
Unarchive a section
}
\examples{
\dontrun{
unarchive_section("12345")
}
}
