% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{update_comment}
\alias{update_comment}
\title{Update a comment}
\usage{
update_comment(
  comment_id,
  content,
  verbose = TRUE,
  token = get_todoist_api_token()
)
}
\arguments{
\item{comment_id}{id of the comment}

\item{content}{new content for the comment}

\item{verbose}{boolean that make the function verbose}

\item{token}{todoist API token}
}
\value{
id of the updated comment (invisible)
}
\description{
Update a comment
}
\examples{
\dontrun{
update_comment("12345", content = "Updated comment")
}
}
