% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{update_filter}
\alias{update_filter}
\title{Update a filter}
\usage{
update_filter(
  filter_id = get_filter_id(filter_name = filter_name, token = token),
  filter_name,
  new_name = NULL,
  query = NULL,
  color = NULL,
  is_favorite = NULL,
  verbose = TRUE,
  token = get_todoist_api_token()
)
}
\arguments{
\item{filter_id}{id of the filter}

\item{filter_name}{name of the filter (for lookup if filter_id not provided)}

\item{new_name}{new name for the filter}

\item{query}{new query string}

\item{color}{new color}

\item{is_favorite}{boolean to mark as favorite}

\item{verbose}{boolean that make the function verbose}

\item{token}{todoist API token}
}
\value{
id of the updated filter (invisible)
}
\description{
Update a filter
}
\examples{
\dontrun{
update_filter(filter_name = "Urgent Today", query = "today & p1 & #Work")
}
}
