% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{update_todoist_api_token}
\alias{update_todoist_api_token}
\title{Update Todoist Api Token}
\usage{
update_todoist_api_token()
}
\value{
nothing, storing your token
}
\description{
Remove the old token and register a new one.
}
