% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{update_workspace}
\alias{update_workspace}
\title{Update a workspace}
\usage{
update_workspace(
  workspace_id,
  name,
  verbose = TRUE,
  token = get_todoist_api_token()
)
}
\arguments{
\item{workspace_id}{id of the workspace}

\item{name}{new name for the workspace}

\item{verbose}{boolean that make the function verbose}

\item{token}{todoist API token}
}
\value{
id of the updated workspace (invisible)
}
\description{
Update a workspace
}
\examples{
\dontrun{
update_workspace("12345", name = "New Workspace Name")
}
}
