% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{users_in_project}
\alias{users_in_project}
\title{Get users in projects}
\usage{
users_in_project(token, project_id)
}
\arguments{
\item{token}{token}

\item{project_id}{projecti id}
}
\value{
dataframe of users in projects
}
\description{
Get users in projects
}
