% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{add_project}
\alias{add_project}
\title{Add a new project}
\usage{
add_project(project_name, verbose = TRUE, token = get_todoist_api_token())
}
\arguments{
\item{project_name}{name of the new project}

\item{verbose}{make it talk}

\item{token}{token}
}
\value{
id of the new project
}
\description{
Add a new project
}
