% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{add_responsible_to_task}
\alias{add_responsible_to_task}
\title{Add responsible to a task}
\usage{
add_responsible_to_task(
  project_id,
  add_responsible,
  task,
  verbose = FALSE,
  token = get_todoist_api_token()
)
}
\arguments{
\item{project_id}{id of project}

\item{add_responsible}{add someone to this task with mail}

\item{task}{the full name of the task}

\item{verbose}{boolean that make the function verbose}

\item{token}{todoist API token}
}
\value{
http request
}
\description{
Add responsible to a task
}
