% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{add_section}
\alias{add_section}
\title{add section}
\usage{
add_section(project_id, section_name, token = get_todoist_api_token())
}
\arguments{
\item{project_id}{project if}

\item{section_name}{section name}

\item{token}{todoist API token}
}
\description{
add section
}
