% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{add_section}
\alias{add_section}
\title{Add section}
\usage{
add_section(
  section_name,
  project_id = get_project_id(project_name = project_name, token = token),
  project_name,
  force = FALSE,
  token = get_todoist_api_token()
)
}
\arguments{
\item{section_name}{section name}

\item{project_id}{id of the project}

\item{project_name}{name of the project}

\item{force}{boolean force section creation even if already exist}

\item{token}{todoist API token}
}
\value{
section id (character)
}
\description{
Add section
}
