% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{add_section}
\alias{add_section}
\title{add section}
\usage{
add_section(
  project_id,
  section_name,
  force = FALSE,
  token = get_todoist_api_token()
)
}
\arguments{
\item{project_id}{id of project}

\item{section_name}{section name}

\item{force}{boolean force section creation even if already exist}

\item{token}{todoist API token}
}
\description{
add section
}
