% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{add_tasks_in_project}
\alias{add_tasks_in_project}
\title{Add a list of tasks}
\usage{
add_tasks_in_project(
  project_id,
  tasks_list,
  try_again = 3,
  time_try_again = 3,
  responsible = NULL,
  due = NULL,
  section_id = NULL,
  token = get_todoist_api_token()
)
}
\arguments{
\item{project_id}{id of project}

\item{tasks_list}{list of tasks}

\item{try_again}{start again the request}

\item{time_try_again}{number of tries}

\item{responsible}{add people in project}

\item{token}{token}

\item{verbose}{make it talk}
}
\value{
id of project (character vector)
}
\description{
Add a list of tasks
}
\examples{
\dontrun{
add_project("my_proj") \%>\%
   add_tasks_in_project(list("First task", "Second task"))
}
}
\seealso{
[add_task_in_project()]
}
