% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{add_tasks_in_project_from_df}
\alias{add_tasks_in_project_from_df}
\title{Add tasks in project}
\usage{
add_tasks_in_project_from_df(
  project_id,
  tasks_as_df,
  verbose = FALSE,
  token = get_todoist_api_token()
)
}
\arguments{
\item{project_id}{id of project}

\item{tasks_as_df}{data.frame of tasks with
c("tasks_list","responsible","due","section_name") names}

\item{verbose}{boolean that make the function verbose}

\item{token}{todoist API token}
}
\value{
id of project (character vector)
}
\description{
Add tasks in project
}
\seealso{
[add_tasks_in_project()]
}
