% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{add_tasks_in_project_from_df}
\alias{add_tasks_in_project_from_df}
\title{Add tasks in project}
\usage{
add_tasks_in_project_from_df(
  project_id = get_project_id(project_name = project_name, token = token),
  tasks_as_df,
  project_name,
  verbose = FALSE,
  token = get_todoist_api_token(),
  update_only = FALSE
)
}
\arguments{
\item{project_id}{id of the project}

\item{tasks_as_df}{data.frame of tasks with
c("tasks_list","responsible","due","section_name") names}

\item{project_name}{name of the project}

\item{verbose}{boolean that make the function verbose}

\item{token}{todoist API token}

\item{update_only}{boolean if tru only update existing (not closed) todo}
}
\value{
id of project (character vector)
}
\description{
Add tasks in project
}
\seealso{
[add_tasks_in_project()]
}
