% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{add_user_in_project}
\alias{add_user_in_project}
\title{Add one user}
\usage{
add_user_in_project(
  project_id,
  mail,
  verbose = TRUE,
  token = get_todoist_api_token()
)
}
\arguments{
\item{project_id}{id of project}

\item{mail}{mail of the user}

\item{verbose}{boolean that make the function verbose}

\item{token}{token}
}
\value{
id of project (character vector)
}
\description{
Add one user
}
\examples{
\dontrun{
get_project_id("test") \%>\% 
   add_user_in_project("jean@mail.fr")
}
}
