% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{add_users_in_project}
\alias{add_users_in_project}
\title{Add a list of users}
\usage{
add_users_in_project(
  project_id,
  list_of_users,
  verbose = TRUE,
  token = get_todoist_api_token()
)
}
\arguments{
\item{project_id}{id of project}

\item{list_of_users}{list of mails}

\item{verbose}{boolean that make the function verbose}

\item{token}{token}
}
\value{
id of project (character vector)
}
\description{
Add a list of users
}
