% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{add_users_in_project}
\alias{add_users_in_project}
\title{Add a list of users}
\usage{
add_users_in_project(
  project_id = get_project_id(project_name = project_name, token = token),
  users_email,
  project_name,
  verbose = TRUE,
  token = get_todoist_api_token()
)
}
\arguments{
\item{project_id}{id of the project}

\item{users_email}{emails of user as character vector}

\item{project_name}{name of the project}

\item{verbose}{boolean that make the function verbose}

\item{token}{token}
}
\value{
id of project (character vector)
}
\description{
Add a list of users
}
