% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{ask_todoist_api_token}
\alias{ask_todoist_api_token}
\title{Pop-up to save the token}
\usage{
ask_todoist_api_token(msg = "Register Todoist Api Token")
}
\arguments{
\item{msg}{message to print in the pop-up}
}
\description{
Pop-up to save the token
}
