% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{call_api}
\alias{call_api}
\title{Call the Sync API}
\usage{
call_api(..., url = TODOIST_SYNC_URL, token = get_todoist_api_token())
}
\arguments{
\item{...}{any params of POST request}

\item{url}{url to call}

\item{token}{todoist API token}
}
\value{
list
}
\description{
Call the Sync API
}
