% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_objects.R
\name{get_all}
\alias{get_all}
\title{Get all objects inside a list}
\usage{
get_all(token = get_todoist_api_token())
}
\arguments{
\item{token}{token}
}
\value{
list of all objects
}
\description{
Collect all the objects in a list.
This allows you to explore your to-do list.
}
\examples{
\dontrun{
# Set API key first
set_todoist_api_token()
# Get all objects
objects <- get_all()
}
}
