% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_objects.R
\name{get_projects}
\alias{get_projects}
\title{List of projects}
\usage{
get_projects(token = get_todoist_api_token())
}
\arguments{
\item{token}{token}
}
\value{
list of all objects
}
\description{
List of projects
}
