% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{get_section_id}
\alias{get_section_id}
\title{get id section}
\usage{
get_section_id(
  project_id = get_project_id(project_name = project_name, token = token),
  project_name,
  section_name,
  token = get_todoist_api_token()
)
}
\arguments{
\item{project_id}{id of the project}

\item{project_name}{name of the project}

\item{section_name}{name of the section}

\item{token}{token}
}
\description{
get id section
}
