% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{get_section_id}
\alias{get_section_id}
\title{Get section id}
\usage{
get_section_id(
  project_id = get_project_id(project_name = project_name, token = token),
  project_name,
  section_name,
  token = get_todoist_api_token(),
  all_section = get_section_from_project(project_id = project_id, token = token)
)
}
\arguments{
\item{project_id}{id of the project}

\item{project_name}{name of the project}

\item{section_name}{name of the section}

\item{token}{token}

\item{all_section}{all_section}
}
\value{
section id (character). Returns 0 if section not found.
}
\description{
Get section id
}
