% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_objects.R
\name{get_tasks_of_project}
\alias{get_tasks_of_project}
\title{List of tasks of project}
\usage{
get_tasks_of_project(project_id, token = get_todoist_api_token())
}
\arguments{
\item{project_id}{id of project}

\item{token}{todoist API token}
}
\value{
list of all tasks
}
\description{
List of tasks of project
}
