% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_users_id}
\alias{get_users_id}
\title{Get users id}
\usage{
get_users_id(mails, token = get_todoist_api_token())
}
\arguments{
\item{mails}{mails of the person}

\item{token}{token}
}
\value{
id of users
}
\description{
Get users id
}
