% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_presta.R
\name{init_forma_admin}
\alias{init_forma_admin}
\title{Init forma admin}
\usage{
init_forma_admin(
  project_id,
  tasks_list = list("recevoir nom/prenom et mail des stagiaires",
    "envoyer test de pr\\\\u00E9\\\\u00E9valuation",
    "mettre les echeances aux taches suivantes", "envoyer convention",
    "v\\\\u00E9rifier retour convention",
    "s'assurer que la logistique du formateur est OK",
    "envoyer convocation aux stagiaires", "upload feuille emmargement",
    "envoyer facture", "verifier que la facture est bien dans le bon projet axonaut",
    "envoyer attestation", "envoyer evaluation a chaud",
    "plannifier l'evaluation a froid", "verifie retour eval a cahud",     
    "verifie retour eval \\\\u00E0 froid",
    "mail retour de qualit\\\\u00E9 + audt besoin"),
  try_again = 3,
  time_try_again = 3,
  verbose = TRUE,
  responsible = NULL,
  token = get_todoist_api_token()
)
}
\arguments{
\item{project_id}{id of project}

\item{tasks_list}{lists of tasks}

\item{try_again}{try again the http request}

\item{time_try_again}{number of tries}

\item{verbose}{make the function verbose}

\item{responsible}{add people in project}

\item{token}{token}
}
\description{
Init forma admin
}
