% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_presta.R
\name{init_forma_manager}
\alias{init_forma_manager}
\title{Title}
\usage{
init_forma_manager(
  project_id,
  tasks_list = list("Envoi du devis + contenu",
    "Assigner au formateur les taches du formateur", "Etre Payé"),
  try_again = 3,
  time_try_again = 3,
  verbose = TRUE,
  responsible = NULL,
  token = get_todoist_api_token()
)
}
\arguments{
\item{token}{}
}
\value{

}
\description{
Title
}
