% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{projects}
\alias{projects}
\alias{proj_get_all}
\alias{proj_get}
\alias{proj_get_by_id}
\alias{proj_add}
\alias{proj_add_quick}
\alias{proj_update_by_name}
\alias{proj_update_by_id}
\alias{proj_delete}
\alias{proj_delete_by_id}
\title{Modify projects on the todoist list}
\usage{
proj_get_all(token = use_token())

proj_get(name = NULL, id = NULL, token = use_token())

proj_get_by_id(id, token = use_token())

proj_add(name, parent_id = NULL, token = use_token())

proj_add_quick(name, parent_id = NULL, token = use_token())

proj_update_by_name(name, new_name, token = use_token())

proj_update_by_id(id, new_name, token = use_token())

proj_delete(name = NULL, id = NULL, token = use_token(),
  prompt = TRUE)

proj_delete_by_id(id, token = use_token())
}
\arguments{
\item{token}{todosit API token}

\item{name}{a vector of strings specifyibng the name of the projects}

\item{id}{a vector of integers speciying project ids}

\item{parent_id}{the project id to nest under}

\item{new_name}{a vector of strings speciying the new name of the projects}
}
\description{
Modify projects on the todoist list
}
