% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todoist-package.R
\docType{package}
\name{rtodoist-package}
\alias{rtodoist}
\alias{rtodoist-package}
\title{rtodoist: Create and Manage Todolist using 'Todoist.com' API}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Allows you to interact with the API of the
    "Todoist" platform.  'Todoist' <https://todoist.com/> provides an
    online task manager service for teams.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ThinkR-open/rtodoist}
  \item Report bugs at \url{https://github.com/ThinkR-open/rtodoist/issues}
}

}
\author{
\strong{Maintainer}: Cervan Girard \email{cervan@thinkr.fr} (\href{https://orcid.org/0000-0002-4816-4624}{ORCID})

Authors:
\itemize{
  \item Vincent Guyader \email{vincent@thinkr.fr} (\href{https://orcid.org/0000-0003-0671-9270}{ORCID})
}

Other contributors:
\itemize{
  \item ThinkR [copyright holder, funder]
}

}
\keyword{internal}
