\name{search_todoist}
\alias{search_todoist}
\title{search_todoist}
\usage{
  search_todoist(query, token = getOption("TodoistToken"))
}
\arguments{
  \item{query}{<what param does>}

  \item{token}{<what param does>}
}
\description{
  <full description>
}
\examples{
\dontrun{
search_todoist("today", token)
}
}

