% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{set_todoist_api_token}
\alias{set_todoist_api_token}
\title{Set todoist api token}
\usage{
set_todoist_api_token(token)
}
\arguments{
\item{token}{todoist api token}
}
\description{
This function use keyring to store your token from your todoist profile. To find your token from todoist website, use \code{\link{open_todoist_website_profile}}
}
