% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todoist-package.R
\docType{package}
\name{todoist-package}
\alias{todoist}
\alias{todoist-package}
\title{todoist: Call API todoist}
\description{
package to call the todoist api.
}
\author{
\strong{Maintainer}: Cervan Girard \email{cervan@thinkr.fr}

Authors:
\itemize{
  \item Vincent Guyader \email{vincent@thinkr.fr} (\href{https://orcid.org/0000-0003-0671-9270}{ORCID})
}

Other contributors:
\itemize{
  \item ThinkR [copyright holder, funder]
}

}
\keyword{internal}
