% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todoist-package.R
\docType{package}
\name{todoist-package}
\alias{todoist}
\alias{todoist-package}
\title{todoist: Call API Todoist}
\description{
Package to call the todoist api.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/VincentGuyader/todoist}
  \item Report bugs at \url{https://github.com/VincentGuyader/todoist/issues}
}

}
\author{
\strong{Maintainer}: Cervan Girard \email{cervan@thinkr.fr}

Authors:
\itemize{
  \item Vincent Guyader \email{vincent@thinkr.fr} (\href{https://orcid.org/0000-0003-0671-9270}{ORCID})
}

Other contributors:
\itemize{
  \item ThinkR [copyright holder, funder]
}

}
\keyword{internal}
